missing static_cast

+++ src/la/dvector.hpp
@@ -369,10 +369,10 @@ allreduce(KokkosDVector<T, ARGS...>& C, const Communicator& comm)
   auto C_ptr = C_h.array().data();
   int m = C_h.map().nrows();
   int n = C_h.map().ncols();
-  if (C_h.array().stride(0) != 1) {
+  if (static_cast<int>(C_h.array().stride(0)) != 1) {
     throw std::runtime_error("allreduce, expected stride(0) == 1");
   }
-  if (C_h.array().stride(1) != m) {
+  if (static_cast<int>(C_h.array().stride(1)) != m) {
     throw std::runtime_error("allreduce, expected stride(1) == ncols");
   }
   comm.allreduce(C_ptr, m * n, mpi_op::sum);
